backend/requirements/base.txt


django
psycopg2
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
pre-commit
asgiref==3.4.1
Django==3.2.5
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
python-dotenv~=0.19.0

djangorestframework~=3.12.4
django-database-view==0.3.0
requests==2.28.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.11
djangorestframework-simplejwt==5.2.2
pyjwt==2.7.0
Pillow==9.5.0
drf_yasg==1.20.0
django-treasuremap

python-environ
django-allauth==0.51.0
django-rest-auth
social-auth-app-django==5.0.0
social-auth-core==4.3.0
django-rest-auth
django-jazzmin==2.6.0
dj-rest-auth==4.0.1
django-rest-passwordreset==1.3.0
django-cors-headers==4.0.0

torch
torchvision
ultralytics
opencv-python
numpy
fastapi
uvicorn